experimentally porting to `provider-interop^0.1`
Some notes for comparison with the test that was already here:

The sample service-provider that was part of the existing test had a hidden dependency - it was expecting the `Service::$value` property to be injected at a later by the client bootstrapping that provider.

Shipping a service-provider that depends on a value that is expected to be injected at a later time, is not possible, because `provider-interop` does not include the "decoration" feature that the previous test was relying on. The provider has to be self-contained and complete - it has to fully bootstrap the dependencies it provides.

The current test therefore only tests for the correct implementation of `ServiceProviderInterface` and `ServiceRegistryInterface`, which in the case of this container-implementation, which is *mutable*, means the `Container` class is implementing both interfaces.

Contrast this with the experimental fork of `mindplay/unbox`, which uses an *immutable* container and a separate container-factory:

https://github.com/mindplay-dk/unbox/tree/provider-interop